% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kd_leaders.R
\name{kd_leaders}
\alias{kd_leaders}
\title{Top Players by Kill-Death Ratio}
\usage{
kd_leaders(df, n = 5)
}
\arguments{
\item{df}{A data frame containing player performance data.}

\item{n}{Number of top players to return.}
}
\value{
A data frame of players ranked by kill-death ratio.
}
\description{
Identifies players with the highest kill-death ratios.
}
\examples{
data(player_stats)
kd_leaders(player_stats, n = 5)

}
