% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_kd_ratio.R
\name{plot_kd_ratio}
\alias{plot_kd_ratio}
\title{Plot Kill-Death Ratio by Player}
\usage{
plot_kd_ratio(df)
}
\arguments{
\item{df}{A data frame containing player performance data.}
}
\value{
A ggplot2 object displaying K/D ratios.
}
\description{
Creates a bar chart showing kill-death ratios for each player.
}
\examples{
data(player_stats)
plot_kd_ratio(player_stats)

}
