% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_player_scores.R
\name{plot_player_scores}
\alias{plot_player_scores}
\title{Plot Player Score Distribution}
\usage{
plot_player_scores(df)
}
\arguments{
\item{df}{A data frame containing player performance data.}
}
\value{
A ggplot2 object displaying score distribution.
}
\description{
Creates a boxplot showing the distribution of player scores.
}
\examples{
data(player_stats)
plot_player_scores(player_stats)

}
