% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_players.R
\name{top_players}
\alias{top_players}
\title{Top Players by Performance}
\usage{
top_players(df, n = 5)
}
\arguments{
\item{df}{A data frame containing player performance data.}

\item{n}{Number of top players to return.}
}
\value{
A data frame of the top n players ranked by score.
}
\description{
Returns the top players ranked by average score.
}
\examples{
data(player_stats)
top_players(player_stats, n = 5)

}
